Extract sigil related method in `sorbet_sigil` module
